requirements.txt
mostlyai
python-dotenv
pandas
requests